Fix wrong parameter type

+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/devices/garmin/fit/messages/FitRecord.java
@@ -136,7 +136,7 @@ public class FitRecord extends RecordData {
             activityPoint.setHeartRate(getHeartRate());
         }
         if (getEnhancedSpeed() != null) {
-            activityPoint.setSpeed(getEnhancedSpeed());
+            activityPoint.setSpeed(getEnhancedSpeed().floatValue());
         }
         return activityPoint;
     }
